feat: extract standalone functionality and introduce plugin-based router configuration
