+++ src/frontend/Cargo.toml
@@ -50,7 +50,6 @@ humantime.workspace = true
 humantime-serde.workspace = true
 lazy_static.workspace = true
 log-query.workspace = true
-log-store.workspace = true
 meta-client.workspace = true
 num_cpus.workspace = true
 opentelemetry-proto.workspace = true
