+++ src/plugins/Cargo.toml
@@ -13,9 +13,11 @@ clap.workspace = true
 cli.workspace = true
 common-base.workspace = true
 common-error.workspace = true
+common-meta.workspace = true
 datanode.workspace = true
 flow.workspace = true
 frontend.workspace = true
 meta-srv.workspace = true
 serde.workspace = true
 snafu.workspace = true
+standalone.workspace = true
