+++ src/plugins/src/standalone.rs
@@ -12,18 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod cli;
-mod datanode;
-mod flownode;
-mod frontend;
-mod meta_srv;
-mod options;
-mod standalone;
+use common_base::Plugins;
+use common_meta::kv_backend::KvBackendRef;
+use standalone::error::Result;
+use standalone::options::StandaloneOptions;
 
-pub use cli::SubCommand;
-pub use datanode::{setup_datanode_plugins, start_datanode_plugins};
-pub use flownode::{setup_flownode_plugins, start_flownode_plugins};
-pub use frontend::{setup_frontend_plugins, start_frontend_plugins};
-pub use meta_srv::{setup_metasrv_plugins, start_metasrv_plugins};
-pub use options::PluginOptions;
-pub use standalone::{setup_standalone_plugins, start_standalone_plugins};
+use crate::options::PluginOptions;
+
+#[allow(unused_variables)]
+#[allow(unused_mut)]
+pub async fn setup_standalone_plugins(
+    plugins: &mut Plugins,
+    plugin_options: &[PluginOptions],
+    standalone_opts: &StandaloneOptions,
+    metadata_kvbackend: KvBackendRef,
+) -> Result<()> {
+    Ok(())
+}
+
+pub async fn start_standalone_plugins(_plugins: Plugins) -> Result<()> {
+    Ok(())
+}
