+++ src/standalone/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "standalone"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+async-trait.workspace = true
+catalog.workspace = true
+client.workspace = true
+common-base.workspace = true
+common-config.workspace = true
+common-error.workspace = true
+common-macro.workspace = true
+common-meta.workspace = true
+common-options.workspace = true
+common-procedure.workspace = true
+common-query.workspace = true
+common-telemetry.workspace = true
+common-time.workspace = true
+common-version.workspace = true
+common-wal.workspace = true
+datanode.workspace = true
+file-engine.workspace = true
+flow.workspace = true
+frontend.workspace = true
+log-store.workspace = true
+mito2.workspace = true
+query.workspace = true
+serde.workspace = true
+servers.workspace = true
+snafu.workspace = true
+store-api.workspace = true
+tokio.workspace = true
