+++ tests-integration/Cargo.toml
@@ -77,6 +77,7 @@ sqlx = { version = "0.8", features = [
     "postgres",
     "chrono",
 ] }
+standalone.workspace = true
 substrait.workspace = true
 table.workspace = true
 tempfile.workspace = true
